show all additional profiles found with GEDS
